src/authorizer/requirements.txt
boto3==1.34.90
botocore==1.34.90
cryptography==42.0.5
jwcrypto==1.5.6
PyJWT==2.8.0
requests==2.31.0